move balance-related screens into balance package
